+++ lib/scss_lint/constants.rb
@@ -1,0 +1,5 @@
+# Global application constants.
+module SCSSLint
+  REPO_URL = 'https://github.com/causes/scss-lint'
+  BUG_REPORT_URL = "#{REPO_URL}/issues"
+end